+++ unit-sh/src/test/java/storageTests/SensorStorageTest.java
@@ -19,7 +19,7 @@ public class SensorStorageTest {
         }
 
         @Override public int readCurrentValue() { return currentValue; }
-        @Override public int getCurrentReading() { return currentValue; }
+        @Override public double getCurrentReading() { return currentValue; }
         @Override public void simulateValue(int value) { this.currentValue = value; updateTimestamp(); }
     }
